BUG: Simple has base typography defined in layout ( trac #7803)
Moved some style declarations from layout to typography so that the cms
would pick them up.
Moved typography tag to the page.ss before layout is included
Changed typography elements to be wrapped in .typography, rather than a
layout tag within the Layout page templates. This will mean that a
bunch of modules have access to the correct type styling